Added audio create and list functionality.

+++ internal/auth/delivery/http/delivery.go
@@ -149,7 +149,7 @@ func (d *delivery) signin(w http.ResponseWriter, r *http.Request) {
 //
 //	@Security		cookieAuth
 func (d *delivery) logout(w http.ResponseWriter, r *http.Request) {
-	userID, ok := r.Context().Value(mw.ContextUserID).(int)
+	userID, ok := r.Context().Value(mw.ContextUserID).(int64)
 	if !ok {
 		pHTTP.HandleError(w, r, pErrors.ErrReadBody)
 		return
@@ -180,7 +180,7 @@ func (d *delivery) logout(w http.ResponseWriter, r *http.Request) {
 }
 
 func (d *delivery) me(w http.ResponseWriter, r *http.Request) {
-	userID, ok := r.Context().Value(mw.ContextUserID).(int)
+	userID, ok := r.Context().Value(mw.ContextUserID).(int64)
 	if !ok {
 		pHTTP.HandleError(w, r, pErrors.ErrReadBody)
 		return
