+++ internal/sessions/repository.go
@@ -3,7 +3,7 @@ package sessions
 import "context"
 
 type Repository interface {
-	Create(ctx context.Context, userID int) (string, error)
-	Get(ctx context.Context, userID int, authToken string) (int, error)
-	Delete(ctx context.Context, userID int, authToken string) error
+	Create(ctx context.Context, userID int64) (string, error)
+	Get(ctx context.Context, userID int64, authToken string) (int64, error)
+	Delete(ctx context.Context, userID int64, authToken string) error
 }
