+++ internal/users/delivery/http/models_easyjson.go
@@ -153,7 +153,7 @@ func easyjsonD2b7633eDecodeGithubComSlavaShagalovProspeechBackendInternalUsersDe
 		}
 		switch key {
 		case "id":
-			out.ID = int(in.Int())
+			out.ID = int64(in.Int64())
 		case "username":
 			out.Username = string(in.String())
 		case "email":
@@ -195,7 +195,7 @@ func easyjsonD2b7633eEncodeGithubComSlavaShagalovProspeechBackendInternalUsersDe
 	{
 		const prefix string = ",\"id\":"
 		out.RawString(prefix[1:])
-		out.Int(int(in.ID))
+		out.Int64(int64(in.ID))
 	}
 	{
 		const prefix string = ",\"username\":"
